requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
mcp-python>=0.1.0
starlette>=0.14.2
python-multipart>=0.0.5
typing-extensions>=4.0.0
